divide simply to files
