+++ src/index.jsx
@@ -1,50 +1,7 @@
 import React, { useState } from 'react';
 import ReactDOM from 'react-dom';
 
-function Counter({ count, onClick }) {
-  return (
-    <button type="button" onClick={onClick}>
-      Click me!
-      (
-      {count}
-      )
-    </button>
-  );
-}
-
-function Button({ children }) {
-  return (
-    <button type="button">
-      {children}
-    </button>
-  );
-}
-
-function Buttons() {
-  return (
-    <p>
-      {[1, 2, 3].map((i) => (
-        <Button key={i}>
-          {i}
-        </Button>
-      ))}
-    </p>
-  );
-}
-
-function Page({ count, onClick }) {
-  return (
-    <div>
-      <p>Hellom world!</p>
-      <p>Hi</p>
-      <Counter
-        count={count}
-        onClick={onClick}
-      />
-      <Buttons />
-    </div>
-  );
-}
+import Page from './components/Page';
 
 function App() {
   const [state, setState] = useState({ count: 0 });
